Reorganize maps feature
